Clamp captions to screen bounds
When using captions on topbar buttons near the right edge of the screen, the captions run offscreen. This change ensures that the captions are always clamped to the bounds of the screen according to `workspace.CurrentCamera.ViewportSize`. 

Currently there is no space between the caption and the edge of the screen, I think this makes better use of screen real-estate, but happy to make any changes, adjust spacing, etc.